# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DamascusTransit API",
    "description": "Real-time transit tracking and fleet management — multi-tenant SaaS edition.\n\n## Authentication\n\nMost read endpoints are public but require an `?operator=<slug>` query parameter.\n1. **POST /api/v1/auth/login** — exchange email/password for a JWT token\n2. Include the token as `Authorization: Bearer <token>` on protected endpoints\n\nRoles: `super_admin` · `admin` · `dispatcher` · `driver` · `viewer`\n\nTokens expire after 24 hours.",
    "contact": {
      "name": "Damascus Transit System",
      "email": "support@damascustransit.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/routes/{route_id}": {
      "get": {
        "tags": [
          "routes"
        ],
        "summary": "Get Route",
        "description": "Get single route details.",
        "operationId": "get_route_api_routes__route_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Route Id"
            }
          },
          {
            "name": "operator",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Operator slug",
              "title": "Operator"
            },
            "description": "Operator slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stops": {
      "get": {
        "tags": [
          "stops"
        ],
        "summary": "List Stops",
        "description": "List all active stops.",
        "operationId": "list_stops_api_stops_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "operator",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Operator slug",
              "title": "Operator"
            },
            "description": "Operator slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StopResponse"
                  },
                  "title": "Response List Stops Api Stops Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stops/nearest": {
      "get": {
        "tags": [
          "stops"
        ],
        "summary": "Find Nearest Stops",
        "description": "Find nearest stops using PostGIS RPC.",
        "operationId": "find_nearest_stops_api_stops_nearest_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "Lon"
            }
          },
          {
            "name": "radius",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 100,
              "default": 1000,
              "title": "Radius"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NearestStop"
                  },
                  "title": "Response Find Nearest Stops Api Stops Nearest Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stops/{stop_id}/eta": {
      "get": {
        "tags": [
          "stops"
        ],
        "summary": "Get Stop Eta",
        "description": "Return upcoming bus arrivals with ETA for a specific stop.\n\nETA is calculated from live vehicle GPS positions using Haversine distance\nand current speed. Falls back to city-average speed (25 km/h) when a\nvehicle is stationary or speed data is unavailable.",
        "operationId": "get_stop_eta_api_stops__stop_id__eta_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "stop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stop Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "description": "Max arrivals to return",
              "default": 3,
              "title": "Limit"
            },
            "description": "Max arrivals to return"
          },
          {
            "name": "operator",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Operator slug",
              "title": "Operator"
            },
            "description": "Operator slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopETAResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/vehicles": {
      "get": {
        "tags": [
          "vehicles"
        ],
        "summary": "List Vehicles",
        "description": "List all active vehicles with latest positions.",
        "operationId": "list_vehicles_api_vehicles_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "operator",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Operator slug",
              "title": "Operator"
            },
            "description": "Operator slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleResponse"
                  },
                  "title": "Response List Vehicles Api Vehicles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/routes/{route_pk}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Update Route",
        "description": "Update a route's business fields (fare, names, timing, active flag).",
        "operationId": "admin_update_route_api_admin_routes__route_pk__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "route_pk",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Route Pk"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ETAArrival": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "vehicle_name": {
            "type": "string",
            "title": "Vehicle Name"
          },
          "vehicle_name_ar": {
            "type": "string",
            "title": "Vehicle Name Ar"
          },
          "route_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Name"
          },
          "route_name_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Name Ar"
          },
          "eta_minutes": {
            "type": "integer",
            "title": "Eta Minutes"
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km"
          },
          "speed_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Kmh"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "estimated"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "vehicle_name",
          "vehicle_name_ar",
          "eta_minutes",
          "distance_km"
        ],
        "title": "ETAArrival"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NearestStop": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "stop_id": {
            "type": "string",
            "title": "Stop Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "name_ar": {
            "type": "string",
            "title": "Name Ar"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "has_shelter": {
            "type": "boolean",
            "title": "Has Shelter"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stop_id",
          "name",
          "name_ar",
          "latitude",
          "longitude",
          "has_shelter"
        ],
        "title": "NearestStop"
      },
...
      "RouteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "route_id": {
            "type": "string",
            "title": "Route Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "name_ar": {
            "type": "string",
            "title": "Name Ar"
          },
          "route_type": {
            "type": "string",
            "title": "Route Type"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          },
          "avg_duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Duration Min"
          },
          "fare_syp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fare Syp"
          },
          "stop_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "route_id",
          "name",
          "name_ar",
          "route_type"
        ],
        "title": "RouteResponse"
      },
      "RouteUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "name_ar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Ar"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^#[0-9a-fA-F]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 500.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          },
          "avg_duration_min": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1440.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Duration Min"
          },
          "fare_syp": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Fare Syp"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "RouteUpdate"
      },
...
      "StopETAResponse": {
        "properties": {
          "stop_id": {
            "type": "string",
            "title": "Stop Id"
          },
          "stop_name": {
            "type": "string",
            "title": "Stop Name"
          },
          "stop_name_ar": {
            "type": "string",
            "title": "Stop Name Ar"
          },
          "arrivals": {
            "items": {
              "$ref": "#/components/schemas/ETAArrival"
            },
            "type": "array",
            "title": "Arrivals"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "stop_id",
          "stop_name",
          "stop_name_ar",
          "arrivals",
          "updated_at"
        ],
        "title": "StopETAResponse"
      },
      "StopResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "stop_id": {
            "type": "string",
            "title": "Stop Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "name_ar": {
            "type": "string",
            "title": "Name Ar"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "has_shelter": {
            "type": "boolean",
            "title": "Has Shelter"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stop_id",
          "name",
          "name_ar",
          "has_shelter",
          "is_active"
        ],
        "title": "StopResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VehicleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "name_ar": {
            "type": "string",
            "title": "Name Ar"
          },
          "vehicle_type": {
            "type": "string",
            "title": "Vehicle Type"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "assigned_route_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Route Id"
          },
          "assigned_driver_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Driver Id"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "speed_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Kmh"
          },
          "occupancy_pct": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupancy Pct"
          },
          "recorded_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded At"
          },
          "approval_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Status"
          },
          "approved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "approval_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Note"
          },
          "driver_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Name"
          },
          "driver_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Email"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vehicle_id",
          "name",
          "name_ar",
          "vehicle_type",
          "capacity",
          "status"
        ],
        "title": "VehicleResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}